GraphQL_task/requirements.txt
aniso8601==9.0.1
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.92.0
graphene==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
idna==3.4
pydantic==1.10.4
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
starlette-graphene==0.2.0
starlette-graphene3==0.6.0
strawberry-graphql==0.156.4
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
